Enhancing Animal Adoption Through Data Science
In the realm of animal rescue, a powerful image can be a lifeline for countless animals awaiting adoption. This project delves into the art of predicting pet photo appeal to enhance their chances of finding loving homes. By harnessing the power of data science and leveraging both images and metadata, I seek to classify pet photos as 'adoptable' or 'not adoptable.' Through the use of Convolutional Neural Networks (CNNs), I aim to automatically discern visual cues that influence engagement. The goal is to equip shelters with actionable insights, ultimately saving lives one image at a time.